refactor(core): extract HttpClient to break circular dependency
- Create standalone HttpClient class with all HTTP request logic
- Update TransactionAPI to depend on HttpClient instead of SapiomClient
- Simplify SapiomClient to delegate to HttpClient

This breaks the circular dependency where SapiomClient imported
TransactionAPI and TransactionAPI imported SapiomClient.
